Moving filed to subdirectories
Switching space delim with comma to match csv format.
