+++ ex3/server/commands/Command.h
@@ -1,4 +1,4 @@
-#include "DefaultIO.h"
+#include "../DefaultIO.h"
 
 #ifndef CLIENT_COMMAND_H
 #define CLIENT_COMMAND_H
