+++ ex3/server/commands/UploadUnclassified.h
@@ -1,6 +1,6 @@
 #include "Command.h"
 #include "vector"
-#include "Iris.h"
+#include "../Iris.h"
 
 #ifndef CLIENT_UPLOADUNCLASSIFIED_H
 #define CLIENT_UPLOADUNCLASSIFIED_H
