+++ ex3/server/metrics/EuclideanDistance.h
